restructure modules dir too
