+++ api/src/opentrons/hardware_control/modules/mod_abc.py
@@ -6,13 +6,9 @@ from pkg_resources import parse_version
 from typing import Dict, Callable, Any, Tuple, Awaitable, Optional
 from opentrons.config import IS_ROBOT, ROBOT_FIRMWARE_DIR
 from opentrons.hardware_control.util import use_or_initialize_loop
-from .types import BundledFirmware
+from .types import BundledFirmware, UploadFunction, InterruptCallback
 
 mod_log = logging.getLogger(__name__)
-
-InterruptCallback = Callable[[str], None]
-UploadFunction = Callable[[str, str, Dict[str, Any]],
-                          Awaitable[Tuple[bool, str]]]
 
 
 class AbstractModule(abc.ABC):
