+++ api/src/opentrons/hardware_control/modules/types.py
@@ -3,6 +3,12 @@ from pathlib import Path
 
 ThermocyclerStep = Dict[str, float]
 
+InterruptCallback = Callable[[str], None]
+
+UploadFunction = Callable[[str, str, Dict[str, Any]],
+                          Awaitable[Tuple[bool, str]]]
+
+ModuleAtPort = namedtuple('ModuleAtPort', ('port', 'name'))
 
 class BundledFirmware(NamedTuple):
     """ Represents an versioned firmware file, generally bundled into the fs"""
@@ -15,3 +21,12 @@ class BundledFirmware(NamedTuple):
 
 class UpdateError(RuntimeError):
     pass
+
+
+class UnsupportedModuleError(Exception):
+    pass
+
+
+class AbsentModuleError(Exception):
+    pass
+
