+++ api/src/opentrons/hardware_control/modules/update.py
@@ -3,9 +3,31 @@ import logging
 import os
 from pathlib import Path
 from glob import glob
-from typing import Any, Dict, Tuple
+from typing import Any, Dict, Tuple, Optional
+from .mod_abc import AbstractModule
 
 log = logging.getLogger(__name__)
+
+
+async def update_firmware(
+        module: AbstractModule,
+        firmware_file: str,
+        loop: Optional[asyncio.AbstractEventLoop]):
+    """ Apply update of given firmware file to given module.
+
+        raises an UpdateError with the reason for the failure.
+    """
+    cls = type(module)
+    flash_port = await module.prep_for_update()
+    kwargs: Dict[str, Any] = {
+        'stdout': asyncio.subprocess.PIPE,
+        'stderr': asyncio.subprocess.PIPE,
+        'loop': loop
+    }
+    successful, res = await cls.bootloader()(flash_port, firmware_file, kwargs)
+    if not successful:
+        log.info(f'Bootloader reponse: {res}')
+        raise types.UpdateError(res)
 
 
 async def find_bootloader_port():
